pyproject.toml
[project]
name = "mini-marketplace"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.1",
    "fastapi[standard]>=0.121.0",
    "psycopg2>=2.9.11",
    "pytest>=8.4.2",
    "pytest-env>=1.2.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
]
